Add auto-update system and installer resources
Implemented a comprehensive auto-update system including UpdateManager, UpdateChecker, UpdateDownloader, UpdateInstaller, and related UI (UpdateWindow). Updated App.xaml.cs to initialize and handle auto-update logic. Added documentation for auto-update, installation, and installer creation. Included installer scripts and manifests for Inno Setup and WiX, and updated .gitignore for installer outputs and artifacts.

+++ ScreenshotHelper.cs
@@ -168,7 +168,7 @@ namespace PrettyScreenSHOT
             {
                 if (curModule != null)
                 {
-                    hookId = SetWindowsHookEx(WH_KEYBOARD_LL, keyboardHook, GetModuleHandle(curModule.ModuleName), 0);
+                    hookId = SetWindowsHookEx(WH_KEYBOARD_LL, keyboardHook, GetModuleHandle(curModule.ModuleName), (uint)0);
                 }
             }
         }
